# app_bi/api/swag/swagger.yaml
definitions:
  github_com_leaf-rain_raindata_app_bi_internal_data_data.ExaFileUploadAndDownload:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      key:
        description: 编号
        type: string
      name:
        description: 文件名
        type: string
      tag:
        description: 文件标签
        type: string
      updatedAt:
        type: string
      url:
        description: 文件地址
        type: string
    type: object
  github_com_leaf-rain_raindata_app_bi_internal_data_data.SysAuthority:
    properties:
      authorityId:
        description: 角色ID
        type: integer
      authorityName:
        description: 角色名
        type: string
      children:
        items:
          $ref: '#/definitions/github_com_leaf-rain_raindata_app_bi_internal_data_data.SysAuthority'
        type: array
      createdAt:
        description: 创建时间
        type: string
      dataAuthorityId:
        items:
          $ref: '#/definitions/github_com_leaf-rain_raindata_app_bi_internal_data_data.SysAuthority'
        type: array
      defaultRouter:
        description: 默认菜单(默认dashboard)
        type: string
      deletedAt:
        type: string
      parentId:
        description: 父角色ID
        type: integer
      updatedAt:
        description: 更新时间
        type: string
    type: object
...
  github_com_leaf-rain_raindata_app_bi_internal_data_dto.CasbinInReceive:
    properties:
      authorityId:
        description: 权限id
        type: integer
      casbinInfos:
        items:
          $ref: '#/definitions/github_com_leaf-rain_raindata_app_bi_internal_data_dto.CasbinInfo'
        type: array
    type: object
  github_com_leaf-rain_raindata_app_bi_internal_data_dto.CasbinInfo:
    properties:
      method:
        description: 方法
        type: string
      path:
        description: 路径
        type: string
    type: object
...
  github_com_leaf-rain_raindata_app_bi_internal_data_dto.PolicyPathResponse:
    properties:
      paths:
        items:
          $ref: '#/definitions/github_com_leaf-rain_raindata_app_bi_internal_data_dto.CasbinInfo'
        type: array
    type: object
...
  github_com_leaf-rain_raindata_app_bi_third_party_rhttp.PageInfo:
    properties:
      keyword:
        description: 关键字
        type: string
      page:
        description: 页码
        type: integer
      pageSize:
        description: 每页大小
        type: integer
    type: object
  github_com_leaf-rain_raindata_app_bi_third_party_rhttp.PageResult:
    properties:
      list: {}
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
  github_com_leaf-rain_raindata_app_bi_third_party_rhttp.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
info:
  contact: {}
paths:
  /authority/createAuthority:
    post:
      consumes:
      - application/json
      parameters:
      - description: 权限id, 权限名, 父角色id
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_leaf-rain_raindata_app_bi_internal_data_data.SysAuthority'
      produces:
      - application/json
      responses:
        "200":
          description: 创建角色,返回包括系统角色详情
          schema:
            allOf:
            - $ref: '#/definitions/github_com_leaf-rain_raindata_app_bi_third_party_rhttp.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_leaf-rain_raindata_app_bi_internal_data_data.SysAuthority'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 创建角色
      tags:
      - Authority
  /authority/deleteAuthority:
    post:
      consumes:
      - application/json
      parameters:
      - description: 删除角色
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_leaf-rain_raindata_app_bi_internal_data_data.SysAuthority'
      produces:
      - application/json
      responses:
        "200":
          description: 删除角色
          schema:
            allOf:
            - $ref: '#/definitions/github_com_leaf-rain_raindata_app_bi_third_party_rhttp.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 删除角色
      tags:
      - Authority
  /authority/getAuthorityList:
    post:
      consumes:
      - application/json
      parameters:
      - description: 页码, 每页大小
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_leaf-rain_raindata_app_bi_third_party_rhttp.PageInfo'
      produces:
      - application/json
      responses:
        "200":
          description: 分页获取角色列表,返回包括列表,总数,页码,每页数量
          schema:
            allOf:
            - $ref: '#/definitions/github_com_leaf-rain_raindata_app_bi_third_party_rhttp.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_leaf-rain_raindata_app_bi_third_party_rhttp.PageResult'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 分页获取角色列表
      tags:
      - Authority
  /authority/setDataAuthority:
    post:
      consumes:
      - application/json
      parameters:
      - description: 设置角色资源权限
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_leaf-rain_raindata_app_bi_internal_data_data.SysAuthority'
      produces:
      - application/json
      responses:
        "200":
          description: 设置角色资源权限
          schema:
            allOf:
            - $ref: '#/definitions/github_com_leaf-rain_raindata_app_bi_third_party_rhttp.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 设置角色资源权限
      tags:
      - Authority
  /authority/updateAuthority:
    post:
      consumes:
      - application/json
      parameters:
      - description: 权限id, 权限名, 父角色id
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_leaf-rain_raindata_app_bi_internal_data_data.SysAuthority'
      produces:
      - application/json
      responses:
        "200":
          description: 更新角色信息,返回包括系统角色详情
          schema:
            allOf:
            - $ref: '#/definitions/github_com_leaf-rain_raindata_app_bi_third_party_rhttp.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_leaf-rain_raindata_app_bi_internal_data_data.SysAuthority'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 更新角色信息
      tags:
      - Authority
...
  /casbin/UpdateCasbin:
    post:
      consumes:
      - application/json
      parameters:
      - description: 权限id, 权限模型列表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_leaf-rain_raindata_app_bi_internal_data_dto.CasbinInReceive'
      produces:
      - application/json
      responses:
        "200":
          description: 更新角色api权限
          schema:
            allOf:
            - $ref: '#/definitions/github_com_leaf-rain_raindata_app_bi_third_party_rhttp.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 更新角色api权限
      tags:
      - Casbin
  /casbin/getPolicyPathByAuthorityId:
    post:
      consumes:
      - application/json
      parameters:
      - description: 权限id, 权限模型列表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_leaf-rain_raindata_app_bi_internal_data_dto.CasbinInReceive'
      produces:
      - application/json
      responses:
        "200":
          description: 获取权限列表,返回包括casbin详情列表
          schema:
            allOf:
            - $ref: '#/definitions/github_com_leaf-rain_raindata_app_bi_third_party_rhttp.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_leaf-rain_raindata_app_bi_internal_data_dto.PolicyPathResponse'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取权限列表
      tags:
      - Casbin
  /fileUploadAndDownload/deleteFile:
    post:
      parameters:
      - description: 传入文件里面id即可
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_leaf-rain_raindata_app_bi_internal_data_data.ExaFileUploadAndDownload'
      produces:
      - application/json
      responses:
        "200":
          description: 删除文件
          schema:
            allOf:
            - $ref: '#/definitions/github_com_leaf-rain_raindata_app_bi_third_party_rhttp.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 删除文件
      tags:
      - ExaFileUploadAndDownload
  /fileUploadAndDownload/getFileList:
    post:
      consumes:
      - application/json
      parameters:
      - description: 页码, 每页大小
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/github_com_leaf-rain_raindata_app_bi_third_party_rhttp.PageInfo'
      produces:
      - application/json
      responses:
        "200":
          description: 分页文件列表,返回包括列表,总数,页码,每页数量
          schema:
            allOf:
            - $ref: '#/definitions/github_com_leaf-rain_raindata_app_bi_third_party_rhttp.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_leaf-rain_raindata_app_bi_third_party_rhttp.PageResult'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 分页文件列表
      tags:
      - ExaFileUploadAndDownload
  /fileUploadAndDownload/upload:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: 上传文件示例
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: 上传文件示例,返回包括文件详情
          schema:
            allOf:
            - $ref: '#/definitions/github_com_leaf-rain_raindata_app_bi_third_party_rhttp.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_leaf-rain_raindata_app_bi_internal_data_data.ExaFileUploadAndDownload'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 上传文件示例
      tags:
      - ExaFileUploadAndDownload
...
swagger: "2.0"